app/streamlit/requirements.txt
streamlit>=1.28
langchain>=0.0.217
openai>=1.2
duckduckgo-search
anthropic>=0.3.0
trubrics>=1.4.3
streamlit-feedback
langchain-community
chromadb
python-decouple
langchain_google_genai
langchain-deepseek
sentence_transformers
watchdog